Refactor subscription and home views, update CourseCouponEntity logic, improve HomeView UI, enhance error handling with CustomErrorWidget, add SubscribtionViewBodyDiscountDetailsSection enhancements, add new widgets DiscountBadge and PriceCard, include new SVG asset for error display

+++ lib/Core/entities/CourseEntities/CourseCouponEntity.dart
@@ -19,8 +19,8 @@ class CourseCouponEntity {
 
   bool get isExpired => DateTime.now().isAfter(expiryDate);
 
-  double applyDiscount(double originalPrice) {
-    if (!isActive || isExpired) return originalPrice;
+  double applyDiscount(int originalPrice) {
+    if (!isActive || isExpired) return originalPrice.toDouble();
     return originalPrice * (1 - discountPercentage / 100);
   }
